requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-http
cachetools==5.3.0
    # via wipac-rest-tools
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
coloredlogs==15.0.1
    # via
    #   wipac-file-catalog (setup.py)
    #   wipac-telemetry
cryptography==39.0.1
    # via pyjwt
deprecated==1.2.13
    # via opentelemetry-api
dnspython==2.3.0
    # via pymongo
googleapis-common-protos==1.56.2
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.51.1
    # via opentelemetry-exporter-jaeger-proto-grpc
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via requests
ldap3==2.9.1
    # via wipac-file-catalog (setup.py)
motor==3.1.1
    # via wipac-file-catalog (setup.py)
opentelemetry-api==1.16.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   wipac-telemetry
opentelemetry-exporter-jaeger==1.16.0
    # via wipac-telemetry
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.16.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-http==1.16.0
    # via wipac-telemetry
opentelemetry-proto==1.16.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.16.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   wipac-telemetry
opentelemetry-semantic-conventions==0.37b0
    # via opentelemetry-sdk
protobuf==3.20.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
    #   wipac-telemetry
pyasn1==0.4.8
    # via ldap3
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via wipac-rest-tools
pymongo==4.3.3
    # via
    #   motor
    #   wipac-file-catalog (setup.py)
pypng==0.20220715.0
    # via qrcode
qrcode==7.4.2
    # via wipac-rest-tools
requests==2.28.2
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-futures
    #   requests-toolbelt
    #   wipac-dev-tools
    #   wipac-file-catalog (setup.py)
    #   wipac-rest-tools
requests-futures==1.0.0
    # via
    #   wipac-file-catalog (setup.py)
    #   wipac-rest-tools
requests-toolbelt==0.10.1
    # via wipac-file-catalog (setup.py)
six==1.16.0
    # via thrift
thrift==0.16.0
    # via opentelemetry-exporter-jaeger-thrift
tornado==6.2
    # via
    #   wipac-file-catalog (setup.py)
    #   wipac-rest-tools
types-cryptography==3.3.23.2
    # via pyjwt
typing-extensions==4.5.0
    # via
    #   opentelemetry-sdk
    #   qrcode
    #   wipac-dev-tools
    #   wipac-telemetry
urllib3==1.26.14
    # via requests
wipac-dev-tools==1.6.12
    # via
    #   wipac-rest-tools
    #   wipac-telemetry
wipac-rest-tools[telemetry]==1.4.12
    # via wipac-file-catalog (setup.py)
wipac-telemetry==0.2.5
    # via wipac-rest-tools
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
